Handle execution not starting timing case
It's possible that execution does not start, which means that the
callback to `executionDidStart` will never be executed. However, the
federatedExtension relies on this callback for lifecycle flow. To work
around this, we also include a callback at the end of the request to
stop timing to avoid illegal state.

This commit also changes all thrown errors in the federated extension to
be logged errors instead, returning null or empty lambdas where
appropriate.